separated webpack to webpack.config.common and webpack.config.local to extend the common configuration, and put webpack.config.common in karma.conf.js.
