+++ karma.conf.js
@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const webpack = require('webpack');
-
+const webpackCommon = require('./webpack.config.common');
 // Karma configuration
 // Generated on Thu Jul 06 2017 08:19:42 GMT+0200 (CEST)
 
@@ -77,63 +77,7 @@ module.exports = function(config) {
     // if true, Karma captures browsers, runs the tests and exits
     singleRun: false,
 
-    webpack: {
-        devtool: 'inline-source-map',
-        resolve: {
-            extensions: ['.ts', '.tsx', '.js']
-        },
-        plugins: [
-            new ExtractTextPlugin('[chunkhash].[name].css'),
-        ],
-        module: {
-          rules: [
-              {
-                  test: /\.css$/,
-                  use: ExtractTextPlugin.extract({
-                      use: 'css-loader'
-                  })
-              },
-              {
-                  test: /\.scss$/,
-                  use: ExtractTextPlugin.extract({
-                      use: [{
-                          loader: 'css-loader'
-                      }, {
-                          loader: 'sass-loader'
-                      }]
-                  })
-              },
-              {
-                  test: /\.js$/,
-                  include: [
-                      path.resolve(__dirname, 'src'),
-                      path.dirname(require.resolve('metabolica'))
-                  ],
-                  loader: 'babel-loader',
-                  query: {
-                      presets: ['es2015', 'stage-0'],
-                      plugins: [
-                          'transform-runtime'
-                      ]
-                  }
-              },
-              {
-                  test: /\.html$/,
-                  loader: 'html-loader'
-              },
-              {
-                  test: /\.(jpe?g|png|svg)$/,
-                  loader: 'file-loader?name=[path][name].[ext]'
-              },
-              { 	test: /\.tsx?$/,
-                  loader: "ts-loader",
-                  include: [
-                      path.resolve(__dirname, 'src')
-                  ]
-              }
-            ]
-        }
-    },
+    webpack: webpackCommon,
 
     // Concurrency level
     // how many browser should be started simultaneous
